+++ src/models/TaxRate.php
@@ -9,6 +9,7 @@ namespace craft\commerce\models;
 
 use Craft;
 use craft\commerce\base\Model;
+use craft\commerce\base\TaxIdValidatorInterface;
 use craft\commerce\Plugin;
 use craft\commerce\records\TaxRate as TaxRateRecord;
 use craft\errors\DeprecationException;
@@ -63,15 +64,10 @@ class TaxRate extends Model
     public bool $removeIncluded = false;
 
     /**
-     * @var bool Whether an included VAT tax amount should be removed from VAT-disqualified subject prices
+     * @var bool Whether an included VAT ID tax amount should be removed from VAT-disqualified subject prices
      * @since 3.4
      */
     public bool $removeVatIncluded = false;
-
-    /**
-     * @var bool Whether this tax rate represents VAT
-     */
-    public bool $isVat = false;
 
     /**
      * @var string The subject to which `$rate` should be applied. Options:
@@ -95,6 +91,11 @@ class TaxRate extends Model
     public ?int $taxZoneId = null;
 
     /**
+     * @var array|null Tax ID Validators
+     */
+    public array $taxIdValidators = [];
+
+    /**
      * @var DateTime|null
      * @since 3.4
      */
@@ -115,6 +116,16 @@ class TaxRate extends Model
      * @var TaxAddressZone|null
      */
     private ?TaxAddressZone $_taxZone = null;
+
+    /**
+     * @inheritdoc
+     */
+    public function attributes(): array
+    {
+        $names = parent::attributes();
+        $names[] = 'isVat'; // TODO remove in Commerce 6.x
+        return $names;
+    }
 
     /**
      * @inheritdoc
@@ -206,6 +217,62 @@ class TaxRate extends Model
     public function getIsEverywhere(): bool
     {
         return !$this->getTaxZone();
+    }
+
+    /**
+     * @return bool
+     * @deprecated in 4.8.0.
+     */
+    public function getIsVat(): bool
+    {
+        Craft::$app->getDeprecator()->log(__METHOD__, 'TaxRate::setIsVat() is deprecated.');
+
+        return $this->hasTaxIdValidators();
+    }
+
+    /**
+     * @return bool
+     * @deprecated in 4.8.0.
+     */
+    public function setIsVat(): void
+    {
+        Craft::$app->getDeprecator()->log(__METHOD__, 'TaxRate::setIsVat() is deprecated.');
+    }
+
+    /**
+     * @return bool
+     * @since 4.8.0
+     */
+    public function hasTaxIdValidators(): bool
+    {
+        return count($this->taxIdValidators) > 0;
+    }
+
+    /**
+     * @param string $className
+     * @return bool
+     * @since 4.8.0
+     */
+    public function hasTaxIdValidator(string $className): bool
+    {
+        return in_array($className, $this->taxIdValidators, true);
+    }
+
+    /**
+     * @return TaxIdValidatorInterface[]
+     * @throws InvalidConfigException
+     */
+    public function getSelectedEnabledTaxIdValidators(): array
+    {
+        $selectedValidators = $this->taxIdValidators;
+        $validators = Plugin::getInstance()->getTaxes()->getTaxIdValidators();
+        $activeValidators = [];
+        foreach ($validators as $validator) {
+            if ($validator::isEnabled() & in_array($validator::class, $selectedValidators)) {
+                $activeValidators[] = $validator;
+            }
+        }
+        return $activeValidators;
     }
 
     /**