+++ src/services/Taxes.php
@@ -9,8 +9,11 @@ namespace craft\commerce\services;
 
 use craft\base\Component;
 use craft\commerce\base\TaxEngineInterface;
+use craft\commerce\base\TaxIdValidatorInterface;
 use craft\commerce\engines\Tax;
 use craft\commerce\events\TaxEngineEvent;
+use craft\commerce\events\TaxIdValidatorsEvent;
+use craft\commerce\taxidvalidators\EuVatIdValidator;
 use yii\base\InvalidConfigException;
 
 /**
@@ -21,6 +24,27 @@ use yii\base\InvalidConfigException;
  */
 class Taxes extends Component implements TaxEngineInterface
 {
+    /**
+     * @event TaxIdValidatorsEvent The event that is raised when tax ID validators are registered.
+     *
+     * Any validator added must be a TaxIdValidatorInterface.
+     *
+     * ```php
+     * use craft\commerce\events\TaxIdValidatorsEvent;
+     * use craft\commerce\services\Taxes;
+     * use yii\base\Event;
+     *
+     * Event::on(
+     *     Taxes::class,
+     *     Taxes::EVENT_REGISTER_TAX_ID_VALIDATORS,
+     *     function(TaxIdValidatorsEvent $event) {
+     *          $event->validators[] = new MyTaxIdValidator();
+     *     }
+     * );
+     * ```
+     */
+    public const EVENT_REGISTER_TAX_ID_VALIDATORS = 'registerTaxIdValidators';
+
     /**
      * @event TaxEngineEvent The event that is triggered when determining the tax engine.
      * @since 3.1
@@ -48,10 +72,45 @@ class Taxes extends Component implements TaxEngineInterface
     public const EVENT_REGISTER_TAX_ENGINE = 'registerTaxEngine';
 
     /**
+     * @var ?TaxEngineInterface $engine The tax engine
+     */
+    private ?TaxEngineInterface $_taxEngine = null;
+
+    /**
+     * @return TaxIdValidatorInterface[]
+     * @throws InvalidConfigException
+     */
+    public function getTaxIdValidators(): array
+    {
+        $validators = [];
+        $validators[] = new EuVatIdValidator();
+
+        $event = new TaxIdValidatorsEvent([
+            'validators' => $validators,
+        ]);
+
+        if ($this->hasEventHandlers(self::EVENT_REGISTER_TAX_ID_VALIDATORS)) {
+            $this->trigger(self::EVENT_REGISTER_TAX_ID_VALIDATORS, $event);
+        }
+
+        foreach ($event->validators as $validator) {
+            if (!$validator instanceof TaxIdValidatorInterface) {
+                throw new InvalidConfigException('Tax ID validator must implement TaxIdValidatorInterface');
+            }
+        }
+
+        return $event->validators;
+    }
+
+    /**
      * Get the current tax engine.
      */
     public function getEngine(): TaxEngineInterface
     {
+        if ($this->_taxEngine !== null) {
+            return $this->_taxEngine;
+        }
+
         $event = new TaxEngineEvent(['engine' => new Tax()]);
 
         // Only allow third party tax engines for PRO edition
@@ -64,7 +123,9 @@ class Taxes extends Component implements TaxEngineInterface
             throw new InvalidConfigException('No tax engine has been registered.');
         }
 
-        return $event->engine;
+        $this->_taxEngine = $event->engine;
+
+        return $this->_taxEngine;
     }
 
     /**